requirements.txt
streamlit
pandas
seaborn
matplotlib
scikit-learn
python-docx

